+++ ft_putptr.c
@@ -6,7 +6,7 @@
 /*   By: dswietoc <dswietoc@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/21 19:33:31 by dswietoc          #+#    #+#             */
-/*   Updated: 2026/07/25 14:17:28 by dswietoc         ###   ########.fr       */
+/*   Updated: 2026/08/01 11:30:13 by dswietoc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 int	ft_putptr(void *ptr)
 {
-	unsigned long	addr;
+	unsigned long long	addr;
 	int		strlen;
 
 	strlen = 0;
-	addr = (unsigned long)ptr;
+	addr = (unsigned long long)ptr;
 	if (!ptr)
 		return (ft_putstr("(nil)"));
 	strlen += ft_putstr("0x");
